+++ cucumber-tsflow/src/cli/run.ts
@@ -11,10 +11,14 @@ function logErrorMessageAndExit(message: string): void {
 }
 
 export default async function run(): Promise<void> {
-	validateNodeEngineVersion(process.version, (error: any) => {
-		console.error(error);
-		process.exit(1);
-	});
+	validateNodeEngineVersion(
+		process.version,
+		(error: any) => {
+			console.error(error);
+			process.exit(1);
+		},
+		console.warn
+	);
 
 	const cli = new Cli({
 		argv: process.argv,