+++ cucumber-tsflow/src/cucumber/binding-decorator.ts
@@ -11,24 +11,14 @@ import {
 	When,
 	World
 } from '@cucumber/cucumber';
-import * as messages from '@cucumber/messages';
 
-import _ from 'underscore';
-import logger from '../utils/logger';
 import { BindingRegistry, DEFAULT_TAG } from './binding-registry';
-import { ManagedScenarioContext } from './managed-scenario-context';
 import { StepBinding, StepBindingFlags } from '../types/step-binding';
 import { ContextType, StepPattern, TypeDecorator } from '../types/types';
-
+import messageCollector from './message-collector';
 interface WritableWorld extends World {
 	[key: string]: any;
 }
-
-/**
- * The property name of the current scenario context that will be attached to the Cucumber
- * world object.
- */
-const SCENARIO_CONTEXT_SLOTNAME = '__SCENARIO_CONTEXT';
 
 /**
  * A set of step patterns that have been registered with Cucumber.
@@ -52,7 +42,6 @@ const stepPatternRegistrations = new Map<StepPattern, StepBindingFlags>();
  */
 export function binding(requiredContextTypes?: ContextType[]): TypeDecorator {
 	return <T>(target: { new (...args: any[]): T }) => {
-		ensureSystemBindings();
 		const bindingRegistry = BindingRegistry.instance;
 		bindingRegistry.registerContextTypesForTarget(target.prototype, requiredContextTypes);
 		const allBindings: StepBinding[] = [];
@@ -78,33 +67,6 @@ export function binding(requiredContextTypes?: ContextType[]): TypeDecorator {
 }
 
 /**
- * Ensures that the 'cucumber-tsflow' hooks are bound to Cucumber.
- *
- * @param cucumber The cucumber object.
- *
- * The hooks will only be registered with Cucumber once regardless of which binding invokes the
- * function.
- */
-const ensureSystemBindings = _.once(() => {
-	Before(function (this: WritableWorld, scenario) {
-		logger.trace('Setting up scenario context for scenario:', JSON.stringify(scenario));
-
-		this[SCENARIO_CONTEXT_SLOTNAME] = new ManagedScenarioContext(
-			scenario.pickle?.name ?? '',
-			_.map(scenario.pickle?.tags ?? new Array<messages.PickleTag>(), (tag: messages.PickleTag) => tag?.name ?? '')
-		);
-	});
-
-	After(function (this: WritableWorld) {
-		const scenarioContext = this[SCENARIO_CONTEXT_SLOTNAME] as ManagedScenarioContext;
-
-		if (scenarioContext) {
-			scenarioContext.dispose();
-		}
-	});
-});
-
-/**
  * Binds a step definition to Cucumber.
  *
  * @param stepBinding The [[StepBinding]] that represents a 'given', 'when', or 'then' step definition.
@@ -113,43 +75,46 @@ function bindStepDefinition(stepBinding: StepBinding): void {
 	const bindingFunc = function (this: WritableWorld): any {
 		const bindingRegistry = BindingRegistry.instance;
 
-		const scenarioContext = this[SCENARIO_CONTEXT_SLOTNAME] as ManagedScenarioContext;
-
-		const matchingStepBindings = bindingRegistry.getStepBindings(
-			stepBinding.stepPattern.toString(),
-			scenarioContext.scenarioInfo.tags
-		);
-
-		if (matchingStepBindings.length > 1) {
-			let message = `Ambiguous step definitions for '${matchingStepBindings[0].stepPattern}':\n`;
-
-			matchingStepBindings.forEach(matchingStepBinding => {
-				message =
-					message +
-					`\t\t${String(matchingStepBinding.targetPropertyKey)} (${matchingStepBinding.callsite.toString()})\n`;
-			});
-
-			throw new Error(message);
-		} else if (matchingStepBindings.length === 0) {
-			throw new Error(
-				`Cannot find matched step definition for ${stepBinding.stepPattern.toString()} with tag ${
-					scenarioContext.scenarioInfo.tags
-				} in binding registry`
+		const scenarioContext = messageCollector.getStepScenarioContext(stepBinding);
+		if (scenarioContext) {
+			const matchingStepBindings = bindingRegistry.getStepBindings(
+				stepBinding.stepPattern.toString(),
+				scenarioContext.scenarioInfo.tags
 			);
+
+			if (matchingStepBindings.length > 1) {
+				let message = `Ambiguous step definitions for '${matchingStepBindings[0].stepPattern}':\n`;
+
+				matchingStepBindings.forEach(matchingStepBinding => {
+					message =
+						message +
+						`\t\t${String(matchingStepBinding.targetPropertyKey)} (${matchingStepBinding.callsite.toString()})\n`;
+				});
+
+				throw new Error(message);
+			} else if (matchingStepBindings.length === 0) {
+				throw new Error(
+					`Cannot find matched step definition for ${stepBinding.stepPattern.toString()} with tag ${
+						scenarioContext.scenarioInfo.tags
+					} in binding registry`
+				);
+			}
+
+			const contextTypes = bindingRegistry.getContextTypesForTarget(matchingStepBindings[0].targetPrototype);
+			const bindingObject = scenarioContext.getOrActivateBindingClass(
+				matchingStepBindings[0].targetPrototype,
+				contextTypes
+			);
+
+			bindingObject._worldObj = this;
+
+			return (bindingObject[matchingStepBindings[0].targetPropertyKey] as () => void).apply(
+				bindingObject,
+				arguments as any
+			);
+		} else {
+			throw new Error('Unable to find the Scenario Context for Hook!');
 		}
-
-		const contextTypes = bindingRegistry.getContextTypesForTarget(matchingStepBindings[0].targetPrototype);
-		const bindingObject = scenarioContext.getOrActivateBindingClass(
-			matchingStepBindings[0].targetPrototype,
-			contextTypes
-		);
-
-		bindingObject._worldObj = this;
-
-		return (bindingObject[matchingStepBindings[0].targetPropertyKey] as () => void).apply(
-			bindingObject,
-			arguments as any
-		);
 	};
 
 	Object.defineProperty(bindingFunc, 'length', {
@@ -183,19 +148,22 @@ function bindHook(stepBinding: StepBinding): void {
 			? function (this: any): any {
 					return stepBinding.targetPrototype[stepBinding.targetPropertyKey].apply() as () => void;
 			  }
-			: function (this: any): any {
-					const scenarioContext = this[SCENARIO_CONTEXT_SLOTNAME] as ManagedScenarioContext;
-					const contextTypes = BindingRegistry.instance.getContextTypesForTarget(stepBinding.targetPrototype);
-					const bindingObject = scenarioContext.getOrActivateBindingClass(stepBinding.targetPrototype, contextTypes);
-					bindingObject._worldObj = this;
-					return (bindingObject[stepBinding.targetPropertyKey] as () => void).apply(bindingObject, arguments as any);
+			: function (this: any, arg: any): any {
+					const scenarioContext = messageCollector.getHookScenarioContext(arg);
+					if (scenarioContext) {
+						const contextTypes = BindingRegistry.instance.getContextTypesForTarget(stepBinding.targetPrototype);
+						const bindingObject = scenarioContext.getOrActivateBindingClass(stepBinding.targetPrototype, contextTypes);
+						bindingObject._worldObj = this;
+						return (bindingObject[stepBinding.targetPropertyKey] as () => void).apply(bindingObject, arguments as any);
+					} else {
+						throw new Error('Unable to find the Scenario Context for Hook!');
+					}
 			  };
-
 	Object.defineProperty(bindingFunc, 'length', {
 		value: stepBinding.argsLength
 	});
 
-	const tags = stepBinding.tag === DEFAULT_TAG ? undefined : stepBinding.tag;
+	const tags = stepBinding.tags === DEFAULT_TAG ? undefined : stepBinding.tags;
 
 	switch (stepBinding.bindingType) {
 		case StepBindingFlags.beforeAll: {