+++ cucumber-tsflow/src/transpilers/vue-sfc/script.ts
@@ -6,26 +6,22 @@ import { resolveTemplateCompilerOptions } from './template';
 const clientCache = new WeakMap<SFCDescriptor, SFCScriptBlock | null>();
 const ssrCache = new WeakMap<SFCDescriptor, SFCScriptBlock | null>();
 
-export function getResolvedScript(descriptor: SFCDescriptor, ssr: boolean): SFCScriptBlock | null | undefined {
+export const getResolvedScript = (descriptor: SFCDescriptor, ssr: boolean): SFCScriptBlock | null | undefined => {
 	return (ssr ? ssrCache : clientCache).get(descriptor);
-}
-
-export function setResolvedScript(descriptor: SFCDescriptor, script: SFCScriptBlock, ssr: boolean): void {
-	(ssr ? ssrCache : clientCache).set(descriptor, script);
-}
+};
 
 // Check if we can use compile template as inlined render function
 // inside <script setup>. This can only be done for build because
 // inlined template cannot be individually hot updated.
-export function isUseInlineTemplate(descriptor: SFCDescriptor, isProd: boolean): boolean {
+export const isUseInlineTemplate = (descriptor: SFCDescriptor, isProd: boolean): boolean => {
 	return isProd && !!descriptor.scriptSetup && !descriptor.template?.src;
-}
+};
 
-export function resolveScript(
+export const resolveScript = (
 	descriptor: SFCDescriptor,
 	options: ResolvedOptions,
 	ssr: boolean
-): SFCScriptBlock | null {
+): SFCScriptBlock | null => {
 	if (!descriptor.script && !descriptor.scriptSetup) {
 		return null;
 	}
@@ -50,5 +46,5 @@ export function resolveScript(
 
 	cacheToUse.set(descriptor, resolved);
 	return resolved;
-}
+};