+++ CHANGELOG.md
@@ -8,21 +8,28 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Please see [CONTRIBUTING.md](https://github.com/LynxWall/cucumber-js-tsflow/blob/master/CONTRIBUTE.md) on how to contribute to cucumber-tsflow.
 
 ## [5.0.1]
+### Changed
+- Removed ManagedScenarioContext from World object along with Before and After hooks used to manage it.
+
 ### Added
+- Transpiler support using typescript or esbuild
+- Vue transform support based on the vite/vue-cucumber plugin
+- Cucumber message collector to manage the ManagedScenarioContext object
+- RegEx matcher to match step expressions with feature step text. Supports all cucumber expressions along with regular expressions.
 - BeforeStep and AfterStep hooks from cucumber with tests
 
 ### Fixed
 - Parameter definitions for different hook functions to be consistent with cucumber
-
+- Step tags to support same functionality as Cucumber hook tags
 
 ## [5.0.0]
 ### Changed
 - BREAKING CHANGE! Renamed JavaScript components used for ts-node and Vue initialization along with snippet and behave formatters.
-- Updated to @cucumber/cucumber version 8.0.0-rc.3
+- Updated to @cucumber/cucumber version 8.0.0
 
 ### Added
 - Implemented CLI and Cucumber Test runner in cucumber-tsflow using cucumber API and libraries
-- Extended cucumber options to add environment (node or vuedom) and debugFile support
+- Extended cucumber options to add environment (node or vue) and debugFile support
 - Cucumber code library update so that summary and reports provide file name and line number of actual test and not location of binding in cucumber-tsflow
 - Added support to dubg individual features associated with step file open in editor
 - Boolean type for cucumber expressions